requirements.txt
sentence-transformers
python-docx
google-generativeai
numpy
torch
transformers
langchain-text-splitters
python-dotenv
qdrant-client
streamlit
pydantic